+++ src/vue3/index.js
@@ -1,3 +1,7 @@
-﻿import ThreeTilesViewer from './ThreeTilesViewer.vue';
+import ThreeTilesViewer from './ThreeTilesViewer.vue';
+
+ThreeTilesViewer.install = function(app) {
+  app.component(ThreeTilesViewer.name, ThreeTilesViewer);
+};
 
 export default ThreeTilesViewer;